requirements.txt
adium-theme-ubuntu==0.3.4
CDApplet==1.0
CDBashApplet==1.0
certifi==2017.11.5
chardet==3.0.4
cryptography==1.7.1
duplicity==0.7.6
enum34==1.1.6
future==0.16.0
idna==2.6
ipaddress==1.0.17
keyring==10.3.1
keyrings.alt==2.2
lockfile==0.12.2
ndg-httpsclient==0.4.2
pyasn1==0.1.9
pycrypto==2.6.1
pygobject==3.22.0
pyOpenSSL==16.2.0
python-telegram-bot==9.0.0
pyxdg==0.25
requests==2.18.4
SecretStorage==2.3.1
shadowsocks==2.8.2
six==1.10.0
unity-lens-photos==1.0
urllib3==1.22
virtualenv==15.1.0
wicd==1.7.4
wxPython==3.0.2.0
wxPython-common==3.0.2.0
